removed warnings from c++ build process
git-svn-id: http://tradelink.googlecode.com/svn/trunk@693 ccc6a032-7b47-0410-9fd8-1f29c74ce1d2

+++ BrokerServers/TradeLibFast/TLMarketBasket.cpp
@@ -22,9 +22,9 @@ namespace TradeLibFast
 		return m;
 	}
 
-	int TLMarketBasket::Count()
+	unsigned int TLMarketBasket::Count()
 	{
-		return _secs.size();
+		return (unsigned int)_secs.size();
 	}
 
 	TLMarketBasket TLMarketBasket::Deserialize(CString msg)
