+++ BrokerServers/TradeLibFast/TLMarketBasket.h
@@ -13,13 +13,13 @@ namespace TradeLibFast
 		~TLMarketBasket(void);
 		CString Serialize();
 		static TLMarketBasket Deserialize(CString msg);
-		int Count();
-		TLSecurity operator[](int index) { return _secs[index]; }
+		unsigned int Count();
+		TLSecurity operator[](unsigned int index) { return _secs[index]; }
 		void Add(TLSecurity sec) { _secs.push_back(sec); }
 		void Add(CString symbol) { _secs.push_back(TLSecurity::Deserialize(symbol)); }
 		void Add(TLMarketBasket basket)
 		{
-			for (size_t i = 0; i<basket.Count(); i++)
+			for (unsigned int i = 0; i<basket.Count(); i++)
 				Add(basket[i]);
 		}
